+++ src/frontend/packages/lib_video/src/components/common/VideoWidgetProvider/FoldableItem/index.spec.tsx
@@ -84,6 +84,7 @@ describe('<FoldableItem />', () => {
     expect(mockSetInfoWidgetModal).toHaveBeenCalledWith({
       title: genericTitle,
       text: genericInfoText,
+      refWidget: expect.any(HTMLDivElement),
     });
   });
 });