+++ src/frontend/packages/lib_video/src/components/common/VideoWidgetProvider/WidgetsContainer/InfoModal/index.tsx
@@ -2,7 +2,7 @@ import { Box, Button, Layer, ResponsiveContext, Text } from 'grommet';
 import { normalizeColor } from 'grommet/utils';
 import { theme } from 'lib-common';
 import { RoundCrossSVG } from 'lib-components';
-import React from 'react';
+import React, { useCallback, useEffect } from 'react';
 import styled from 'styled-components';
 
 const StyledTitleText = styled(Text)`
@@ -17,20 +17,59 @@ interface InfoModalProps {
   text: string;
   title: string;
   onModalClose: () => void;
+  refWidget?: HTMLDivElement | null;
 }
 
-export const InfoModal = ({ text, title, onModalClose }: InfoModalProps) => {
+export const InfoModal = ({
+  text,
+  title,
+  onModalClose,
+  refWidget,
+}: InfoModalProps) => {
   const size = React.useContext(ResponsiveContext);
+  const positionAbove = 200; // px
+  const isFirefox = navigator.userAgent.toLowerCase().indexOf('firefox') > -1;
+
+  useEffect(() => {
+    if (!isFirefox && refWidget) {
+      const timeout = setTimeout(() => {
+        refWidget.scrollIntoView({ block: 'center', behavior: 'smooth' });
+      }, 50);
+
+      return () => clearTimeout(timeout);
+    }
+
+    return () => null;
+  }, [refWidget, isFirefox]);
+
+  const onClose = useCallback(() => {
+    onModalClose();
+
+    if (refWidget) {
+      setTimeout(() => {
+        refWidget.scrollIntoView({ block: 'center', behavior: 'smooth' });
+      }, 200);
+    }
+  }, [onModalClose, refWidget]);
 
   return (
     <Layer
-      onEsc={onModalClose}
-      onClickOutside={onModalClose}
+      onEsc={onClose}
+      onClickOutside={onClose}
       responsive={false}
       style={{
         width: size === 'small' ? '95%' : '500px',
         border: `1px solid ${normalizeColor('blue-active', theme)}`,
+        marginTop: `${
+          refWidget?.offsetTop
+            ? isFirefox
+              ? positionAbove
+              : refWidget?.offsetTop - positionAbove
+            : 0
+        }px`,
       }}
+      position={refWidget?.offsetTop ? 'top' : 'center'}
+      data-testid="info-modal"
     >
       <Box background="bg-info" direction="column" round="6px">
         <Box
@@ -38,7 +77,7 @@ export const InfoModal = ({ text, title, onModalClose }: InfoModalProps) => {
           pad={{ horizontal: 'small', top: 'small' }}
         >
           <Button
-            onClick={onModalClose}
+            onClick={onClose}
             plain
             style={{ display: 'block', padding: 0 }}
           >